fix remove url

+++ lib/tasks/remove.rake
@@ -1,8 +1,11 @@
 namespace :remove do
   desc "TODO"
   task url: :environment do
-    Tweet.where("text like '%http%'").limit(10000).each do |tweet|
-      tweet.destroy
+    loop do
+      return unless Tweet.where("text like '%http%'").exists?
+      Tweet.where("text like '%http%'").limit(10000).each do |tweet|
+        tweet.destroy
+      end
     end
   end
 end
